+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from "./style";
 export * from "./utility";
 export * from "./currency";
-export * from "./tailwind";
 export * from "./string";
 export * from "./date";
 export * from "./breakpoint";